Rename DetectionFrameDeducer to DetectionDeducer, make them responsible for unpacking the container
