+++ src/main/java/io/leonis/subra/ipc/serialization/protobuf/vision/DetectionDeducer.java
@@ -1,7 +1,7 @@
 package io.leonis.subra.ipc.serialization.protobuf.vision;
 
 import io.leonis.subra.game.data.*;
-import io.leonis.subra.ipc.serialization.protobuf.vision.DetectionFrameDeducer.DetectionFrame;
+import io.leonis.subra.ipc.serialization.protobuf.vision.DetectionDeducer.DetectionFrame;
 import io.leonis.zosma.game.engine.Deducer;
 import java.util.Set;
 import java.util.stream.*;
@@ -9,21 +9,24 @@ import lombok.Value;
 import org.reactivestreams.Publisher;
 import org.robocup.ssl.Detection;
 import org.robocup.ssl.Detection.DetectionRobot;
+import org.robocup.ssl.Wrapper.WrapperPacket;
 import reactor.core.publisher.Flux;
 
 /**
- * The Class DetectionFrameDeducer.
+ * The Class DetectionDeducer.
  *
  * This class represents a {@link Deducer} of {@link Detection.DetectionFrame} to {@link DetectionFrame}.
  *
  * @author Rimon Oz
  */
-public class DetectionFrameDeducer implements Deducer<Detection.DetectionFrame, DetectionFrame> {
+public class DetectionDeducer<I extends WrapperPacketSupplier>
+    implements Deducer<I, DetectionFrame> {
   @Override
-  public Publisher<DetectionFrame> apply(
-      final Publisher<Detection.DetectionFrame> detectionFramePublisher
-  ) {
-    return Flux.from(detectionFramePublisher)
+  public Publisher<DetectionFrame> apply(final Publisher<I> dataPublisher) {
+    return Flux.from(dataPublisher)
+        .map(WrapperPacketSupplier::getWrapperPacket)
+        .filter(WrapperPacket::hasDetection)
+        .map(WrapperPacket::getDetection)
         .map(detectionFrame ->
             new DetectionFrame(
                 detectionFrame.getBallsList().stream()
